rename YammerMetricUtils to ServerTestUtils, mv to util package and add yammerMetric value
fixup
