+++ server/src/test/java/org/apache/kafka/server/share/fetch/ShareFetchTestUtils.java
@@ -126,23 +126,5 @@ public class ShareFetchTestUtils {
         );
     }
 
-    /**
-     * Fetch the gauge value from the yammer metrics.
-     *
-     * @param name The name of the metric.
-     * @return The gauge value as a number.
-     */
-    public static Number yammerMetricValue(String name) {
-        try {
-            Gauge gauge = (Gauge) KafkaYammerMetrics.defaultRegistry().allMetrics().entrySet().stream()
-                .filter(e -> e.getKey().getMBeanName().contains(name))
-                .findFirst()
-                .orElseThrow()
-                .getValue();
-            return (Number) gauge.value();
-        } catch (Exception e) {
-            return 0;
-        }
-    }
 
 }
