+++ server/src/test/java/org/apache/kafka/server/share/session/ShareSessionCacheTest.java
@@ -18,7 +18,6 @@ package org.apache.kafka.server.share.session;
 
 import org.apache.kafka.common.Uuid;
 import org.apache.kafka.common.utils.internals.ImplicitLinkedHashCollection;
-import org.apache.kafka.server.metrics.YammerMetricsTestUtils;
 import org.apache.kafka.server.share.CachedSharePartition;
 import org.apache.kafka.server.share.ShareGroupListener;
 import org.apache.kafka.test.TestUtils;
@@ -30,7 +29,8 @@ import org.mockito.Mockito;
 import java.util.Iterator;
 import java.util.List;
 
-import static org.apache.kafka.server.share.fetch.ShareFetchTestUtils.yammerMetricValue;
+import static org.apache.kafka.server.util.ServerTestUtils.clearYammerMetrics;
+import static org.apache.kafka.server.util.ServerTestUtils.yammerMetricValue;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -41,7 +41,7 @@ public class ShareSessionCacheTest {
 
     @BeforeEach
     public void setUp() {
-        YammerMetricsTestUtils.clearYammerMetrics();
+        clearYammerMetrics();
     }
 
     @Test
