+++ storage/src/test/java/org/apache/kafka/server/purgatory/DelayedRemoteListOffsetsTest.java
@@ -25,8 +25,8 @@ import org.apache.kafka.common.protocol.Errors;
 import org.apache.kafka.common.record.internal.FileRecords;
 import org.apache.kafka.common.requests.ListOffsetsResponse;
 import org.apache.kafka.server.metrics.KafkaYammerMetrics;
-import org.apache.kafka.server.metrics.YammerMetricsTestUtils;
 import org.apache.kafka.server.util.timer.MockTimer;
+import org.apache.kafka.server.util.ServerTestUtils;
 import org.apache.kafka.storage.internals.log.AsyncOffsetReadFutureHolder;
 import org.apache.kafka.storage.internals.log.OffsetResultHolder;
 
@@ -66,7 +66,7 @@ public class DelayedRemoteListOffsetsTest {
     @AfterEach
     public void afterEach() throws Exception {
         purgatory.shutdown();
-        YammerMetricsTestUtils.clearYammerMetrics();
+        ServerTestUtils.clearYammerMetrics();
     }
 
     @Test
